Refactor toolbars out of Viewer
